WEBUI-265: extract document service from ftest framework
